refactor: rowClickHandler 코드 리팩토링 및 불필요 atoms 제거
