+++ Client/src/recoil/stockInfo/atoms.ts
@@ -15,8 +15,3 @@ export const currentPriceState = atom<number>({
   key: 'currentPrice',
   default: 0,
 });
-
-export const selectedStockAtom = atom<string>({
-  key: 'selectedStock',
-  default: '',
-});
